requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11.5
django-extensions==1.9.0
djangorestframework==3.6.4
Faker==0.7.3
graphviz==0.8
idna==2.6
mixer==5.6.6
py==1.4.34
pygraphviz==1.3.1
pytest==3.2.1
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
six==1.10.0
urllib3==1.22
xlrd==1.1.0
